+++ src/main/java/com/fundatec/vinilemess/pokedex/service/IPokemonIntegrationService.java
@@ -3,7 +3,7 @@ package com.fundatec.vinilemess.pokedex.service;
 import com.fundatec.vinilemess.pokedex.service.external.PokemonResponse;
 
 public interface IPokemonIntegrationService {
-    PokemonResponse getPokemonResponseById(Long id);
+    PokemonResponse getPokemonResponseById(Integer id);
     PokemonResponse getPokemonResponseByName(String name);
 
     boolean validatePokemonExistenceByName(String name);
